streamline refactor (3)
